Crate artemis_codegen

Crate for internal use by other graphql-client crates, for code generation.

It is not meant to be used directly by users of the library.

Modules

deprecation
Deprecation-related code

normalization
Normalization-related code

schema
Contains the Schema type and its implementation.

wasm

Structs

GraphQLClientCodegenOptions
Used to configure code generation.

Enums

CodegenError
An error that happened during code generation

CodegenMode
Which context is this code generation effort taking place.

Functions

generate_module_token_stream
Generates Rust code given a query document, a schema and options.

generate_root_token_stream
a