Struct artemis_codegen::GraphQLClientCodegenOptions

pub struct GraphQLClientCodegenOptions {
    pub mode: CodegenMode,
    pub operation_name: Option<String>,
    pub struct_name: Option<String>,
    pub include_query_info: bool,
    pub wasm_bindgen: bool,
    // some fields omitted
}

Used to configure code generation.

Fields
mode: CodegenMode

Which context is this code generation effort taking place.

operation_name: Option<String>

Name of the operation we want to generate code for. If it does not match, we use all queries.

struct_name: Option<String>

The name of implemention target struct.

include_query_info: bool

Include query info? (Needed for things like normalized caching)

wasm_bindgen: bool

Include wasm_bindgen utils? Requires wasm-bindgen and wasm-typescript-definition

Methods

impl GraphQLClientCodegenOptions

pub fn new(mode: CodegenMode) -> GraphQLClientCodegenOptions

Creates an empty options object with default params. It probably wants to be configured.

pub fn set_query_file(&mut self, path: PathBuf)

A path to a file to include in the module to force Cargo to take into account changes in the query files when recompiling.

pub fn variables_derives(&self) -> Option<&str>

Comma-separated list of additional traits we want to derive for variables.

pub fn set_variables_derives(&mut self, variables_derives: String)

Comma-separated list of additional traits we want to derive for variables.

pub fn response_derives(&self) -> Option<&str>

Comma-separated list of additional traits we want to derive for responses.

pub fn set_response_derives(&mut self, response_derives: String)

Comma-separated list of additional traits we want to derive for responses.

pub fn set_deprecation_strategy(
&mut self,
deprecation_strategy: DeprecationStrategy
)

The deprecation strategy to adopt.

pub fn set_module_visibility(&mut self, visibility: Visibility)

Target module visibility.

pub fn set_struct_name(&mut self, struct_name: String)

The name of implemention target struct.

pub fn set_operation_name(&mut self, operation_name: String)

Name of the operation we want to generate code for. If none is selected, it means all operations.

pub fn schema_file(&self) -> Option<&Path>

A path to a file to include in the module to force Cargo to take into account changes in the schema files when recompiling.

pub fn query_file(&self) -> Option<&Path>

A path to a file to include in the module to force Cargo to take into account changes in the query files when recompiling.

pub fn set_struct_ident(&mut self, ident: Ident)

The identifier to use when referring to the struct implementing GraphQLQuery, if any.

pub fn struct_ident(&self) -> Option<&Ident>

The identifier to use when referring to the struct implementing GraphQLQuery, if any.

pub fn set_normalization(&mut self, norm: Normalization)

Set the normalization mode for the generated code.

pub fn normalization(&self) -> Normalization

The normalization mode for the generated code.